# readium-api/utils/swagger/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Readium API",
    "description": "Handsome backend devs"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
...
  "paths": {
...
    "/drafts/": {
      "get": {
        "tags": [
          "Draft"
        ],
        "summary": "Fetch all my drafts",
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Draft"
        ],
        "summary": "Initialize a draft",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/drafts/{id}/cover-image": {
      "get": {
        "tags": [
          "Draft"
        ],
        "summary": "Fetch a draft cover image",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/drafts/{id}": {
      "get": {
        "tags": [
          "Draft"
        ],
        "summary": "Fetch a draft",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Draft"
        ],
        "summary": "Create a draft based on published post",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Draft"
        ],
        "summary": "Endpoint to delete a draft",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/drafts/{id}/diff": {
      "patch": {
        "tags": [
          "Draft"
        ],
        "summary": "Update draft's textEditorContent",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextEditorContent"
              }
            }
          }
        }
      }
    },
    "/drafts\r/{id}": {
      "patch": {
        "tags": [
          "Draft"
        ],
        "summary": "Update a draft",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverImage": {
                    "type": "file"
                  },
                  "title": {
                    "type": "string",
                    "example": "Post Title"
                  },
                  "diff": {
                    "$ref": "#/components/schemas/TextEditorContent"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "tag 1",
                      "tag 2"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "example": "Post description"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drafts\r/{id}/cover-image": {
      "put": {
        "tags": [
          "Draft"
        ],
        "summary": "Update draft's cover image",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverImage": {
                    "type": "file"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{id}/title": {
      "put": {
        "tags": [
          "Draft"
        ],
        "summary": "Update draft's title",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "default": "Post title"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{id}/tags": {
      "put": {
        "tags": [
          "Draft"
        ],
        "summary": "Update draft's tags",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "tag 1",
                      "tag 2"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drafts\r/{id}/description": {
      "put": {
        "tags": [
          "Draft"
        ],
        "summary": "Update draft's description",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "default": "Post's description"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{id}/publish": {
      "put": {
        "tags": [
          "Draft"
        ],
        "summary": "Endpoint to publish the draft",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TextEditorContent": {
        "type": "object",
        "properties": {
          "diff": {
            "type": "object",
            "properties": {
              "ops": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "insert": {
                      "type": "string",
                      "example": "Grey"
                    },
                    "attributes": {
                      "type": "object",
                      "properties": {
                        "color": {
                          "type": "string",
                          "example": "#ccc"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}